# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RideDriveAhead API
  description: |
    Enterprise API specification for the Advance-Scheduled Ride Booking Platform.
    Provides complete endpoints for Riders, Drivers, and Admin Compliance Officers,
    with core focus on Advance-Scheduled trips, upfront fare guarantees, driver fatigue protection,
    and AES-256 KYC document vault management.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Gateway
  - url: https://api.ridedriveahead.com/v1
    description: Production API Gateway
paths:
...
  /drivers/kyc/upload:
    post:
      tags: [Driver]
      summary: Upload KYC document (encrypted into AES-256 vault)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycUploadRequest'
      responses:
        '201':
          description: Document securely vaulted and queued for verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycDocumentDto'
  /drivers/kyc/status:
    get:
      tags: [Driver]
      summary: Check overall KYC verification status
      security:
        - BearerAuth: []
      responses:
        '200':
          description: KYC status details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycStatusResponse'
  /drivers/availability:
    post:
      tags: [Driver]
      summary: Toggle driver online/offline availability
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isOnline]
              properties:
                isOnline:
                  type: boolean
                currentLat:
                  type: number
                currentLng:
                  type: number
      responses:
        '200':
          description: Availability updated
  /drivers/scheduled-pool:
    get:
      tags: [Driver]
      summary: View available advance-scheduled rides to claim
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of unassigned scheduled bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDto'
  /drivers/scheduled-pool/{bookingId}/accept:
    post:
      tags: [Driver]
      summary: Claim and accept an advance-scheduled booking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trip claimed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDto'
  /drivers/gigs:
    get:
      tags: [Driver]
      summary: Marketplace of open driving gigs (hourly chauffeur, events, 1-2 day outstation, cabs)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Open gigs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GigOfferDto'
  /drivers/fatigue-status:
    get:
      tags: [Driver]
      summary: Check driver driving-hours and fatigue safety status
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Fatigue monitoring data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FatigueStatusResponse'
  /drivers/earnings:
    get:
      tags: [Driver]
      summary: Get driver earnings breakdown
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daily, weekly, monthly earnings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverEarningsResponse'

  # -------------------------------------------------------------
  # PAYMENT SERVICE
  # -------------------------------------------------------------
  /payments/process:
    post:
      tags: [Payments]
      summary: Process trip payment (UPI, Card, Wallet)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentReceiptDto'
  /payments/receipt/{bookingId}:
    get:
      tags: [Payments]
      summary: Retrieve itemized trip receipt
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Itemized receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentReceiptDto'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookingDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        riderId:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
        bookingType:
          type: string
          enum: [INSTANT, SCHEDULED]
        status:
          type: string
          enum: [DRAFT, SCHEDULED_CONFIRMED, ASSIGNED, EN_ROUTE, ARRIVED, IN_PROGRESS, COMPLETED, CANCELLED]
        scheduledPickupTime:
          type: string
          format: date-time
        pickupAddress:
          type: string
        dropAddress:
          type: string
        vehicleType:
          type: string
        fareAmount:
          type: number
        otpCode:
          type: string
          example: "4821"
        driverName:
          type: string
        driverPhone:
          type: string
        driverRating:
          type: number
        vehicleModel:
          type: string
        vehiclePlate:
          type: string
        flightNumber:
          type: string
        riderNotes:
          type: string
        createdAt:
          type: string
          format: date-time
...
    KycUploadRequest:
      type: object
      required: [documentType, documentNumber, fileBase64]
      properties:
        documentType:
          type: string
          enum: [AADHAAR, PAN, DRIVING_LICENSE, VEHICLE_REGISTRATION]
        documentNumber:
          type: string
          example: "5482 9182 3019"
        fileBase64:
          type: string
          description: Base64 encoded document image or PDF
    KycDocumentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentType:
          type: string
        maskedNumber:
          type: string
          example: "XXXX-XXXX-3019"
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        vaultDocumentId:
          type: string
        uploadedAt:
          type: string
          format: date-time
    KycStatusResponse:
      type: object
      properties:
        overallStatus:
          type: string
          enum: [NOT_SUBMITTED, PENDING, APPROVED, REJECTED]
        isEligibleToDrive:
          type: boolean
        documents:
          type: array
          items:
            $ref: '#/components/schemas/KycDocumentDto'
    FatigueStatusResponse:
      type: object
      properties:
        drivingHoursToday:
          type: number
          example: 4.8
        maxDailyDrivingHours:
          type: number
          example: 8.0
        remainingHoursAllowed:
          type: number
          example: 3.2
        isRestBreakRecommended:
          type: boolean
          example: false
        lastRestTime:
          type: string
          format: date-time
    DriverEarningsResponse:
      type: object
      properties:
        todayEarnings:
          type: number
          example: 2850.0
        weeklyEarnings:
          type: number
          example: 16400.0
        monthlyEarnings:
          type: number
          example: 68200.0
        todayCompletedTrips:
          type: integer
          example: 6
        scheduledTripsCompleted:
          type: integer
          example: 2
        incentivesEarned:
          type: number
          example: 450.0
    ProcessPaymentRequest:
      type: object
      required: [bookingId, paymentMethod, amount]
      properties:
        bookingId:
          type: string
          format: uuid
        paymentMethod:
          type: string
          enum: [UPI, CARD, WALLET, CASH]
        amount:
          type: number
        upiId:
          type: string
          example: "rider@okhdfcbank"
    PaymentReceiptDto:
      type: object
      properties:
        receiptNumber:
          type: string
          example: "RCP-20260911-8921"
        bookingId:
          type: string
          format: uuid
        baseFare:
          type: number
        distanceFare:
          type: number
        advanceReservationFee:
          type: number
        taxes:
          type: number
        totalAmount:
          type: number
        paymentMethod:
          type: string
        paidAt:
          type: string
          format: date-time
...
    GigOfferDto:
      type: object
      properties:
        gigId:
          type: string
          example: "GIG-101"
        gigType:
          type: string
          enum: [HOURLY_CHAUFFEUR, EVENT_CHAUFFEUR, OUTSTATION_MULTI_DAY, AIRPORT_CAB]
        title:
          type: string
          example: "4-Hour Local Chauffeur"
        carModel:
          type: string
          example: "Customer's Honda City (Automatic AT)"
        scheduledTime:
          type: string
          format: date-time
        pickupAddress:
          type: string
        fareAmount:
          type: number
          example: 549.0
        isClaimed:
          type: boolean